terminate stack python example
The python example for stack terminating is wrong. There is no `terminate()` method in the code.
But there is `delete()`: https://github.com/tutumcloud/python-tutum/blob/staging/tutum/api/base.py#L195

So, this is either typo in the doc, or wrong method name in the code, because for tutum-go and tutum-cli is the method consistently named as "terminate".